Fixed _mass2_batch warnings error, made _mass2_gpu its own file, thank God -EM
